Modified how the callback URL for authorization is built.
The URL is now created via a static method in the callback handler's class, AuthorizeDomainRequest. It takes a HttpServletRequest parameter for portability reasons. It can use the incoming request to determine the domain.

Note: This is only tested in that it will generate a valid string. In order to do a more comprehensive test, a shim to test with is required.

Closes #5