Updated how the configuration file was imported.
Before we were storing the custom configuration options in our own Properties object. This is slightly burdensome to external parties and a little confusing. To get around this, we now update the System Properties. This means that command line properties, default propteries, and custom properties are all imported, in that order.